[OU-FIX] sale_coupon_multiplier_free_product: Link properly coupon programs
Both code_promo_program_id and applied_coupon_ids are empty for this
kind of promotion, so we can't do the search by such fields.

We dig onto the reward product + promotion line to infer the applied
promotion. This code is not perfect, as if you have made the same
promotion several times, only last one will be linked to all the
rewards, but as general solution serves.

TT33870