[FIX] sale_coupon_multiplier_free_product: handle invalid lines
We have to put some logic redundancy here as the main method doesn't have
enough granularity to avoid deleting the lines belonging to the multiplier
programs when the promotions are updated. So the main module expects that the
promotion lines products match with the promotion discount product
(https://git.io/JWpoU) , which is not the approach in this module, where we add
extra lines with the reward products themselves and the proper price tag and
discount. So in this method override, we'll save those correct lines from the
fire via context that the unlink method will properly catch. We also have to
remove the proper invalid lines that wouldn't be detected

TT38806