[FIX] sale_coupon_domain_product_discount: fix error when sending coupon by email
TT41937

When the option to apply discount on specific products is checked,
it expects that there are records in the field where those products
are defined. If this option is checked and there are no products
defined, an error occurs when generating the report that is attached
to the email to send the customer their coupon because it tries to
read the records of these products and they do not exist.
This has been corrected by including this option in the report to
adapt it to this module and not to fail by creating a description
field that will be printed in the report to send the information to
the client.