[FIX] loyalty_partner_applicability: Set default value for partner domain rule
It is necessary to establish the default value, in this case "[ ]" to the
rule_partners_domain field, in the create method for the rules established
in the computed method "_program_type_default_values" that establishes the
default values depending on the type of program, because when creating a
new program, these rules do not have the rule_partners_domain field defined,
as it is not contemplated, its value will be "False", which cannot be
interpreted as a correct domain. To avoid this error, the check is made in
the create method and the correct value is added.

TT44344