[FIX] loyalty_criteria_multi_product: Avoiding cross-data in multi-product promotion criteria
When setting up a multi-product promotion, you should not set a minimum
spend amount because if one rule passes, they will all pass.
To avoid misconfiguration, the field is hidden for promotions
with multi-product criteria. In addition, the rule data is cleared
when the criterion is changed so that the promotion is applied
correctly in case of promotion modifications.