[IMP] website_sale_loyalty_suggestion_wizard: Show full product name
When setting up a multi-gift promotion or even in a simple promotion,
showing the product name without differentiating that it may be a variant,
the customer does not know exactly what gift they are receiving.
To avoid this confusion and also when choosing alternatives in a
multi-gift promotion, the full name is shown with its attributes but
omitting the reference.