[IMP] loyalty_partner_applicability: Improve the applicability of loyalty programs to partners
This change introduces a new mixin to allow the definition of the applicability
of a loyalty program to a partner. This mixin is used in the loyalty program model
to define the applicability of the program to partners. The mixin is also used
in the loyalty rule model to define the applicability of the rule to partners.

With this change, the applicability of a loyalty program to a partner can be
now defined at the program and rule level.