[FIX] sale_loyalty_order_suggestion: do not apply promotions automatically
When there is a promotion of the type next_order_coupons, this generates
a coupon for an order that has not yet been saved and therefore that
coupon does not have the associated order that generated it. Furthermore,
this line is unnecessary since promotions are applied automatically and,
if the order already meets the conditions for a promotion to be applied,
it will still be suggested in the wizard and the products added/required
to apply the selected promotion will also be reported.

For these reasons, the line is removed as it applies redundant and
unnecessary logic that slows down the process and generates useless
coupons.