[IMP] sale_loyalty_order_suggestion_multi_gift: adjust multi-gift rewards when required product = reward product
Before this correction, when the required product and the reward product
were the same and the order met the rule, the paid order line was
incorrectly reduced, resulting in erroneous reward quantities.

Now, if the product is both required and rewarded, the order line
remains unchanged, reducing only the reward quantities if applicable,
as expected in its normal behaviour.

TT58167